[assembly] Assert update package size during assembly
Instead of checking the update package size after assembly is complete,
we now do so when the update package is built. We are able to do this
because of the recent SWD changes that make the update package's size
product-agnostic. We set the budget to 200 KiB to start. The current
size is only 64 KiB.

Bug: b/309475849
Change-Id: I2cb2d1f3301632a5bebf3b09c913262275e903f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972615
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>