[audio] SetActiveChannels set time
This CL adds an out-parameter for the SimpleAudioStream
method ChangeActiveChannels, so that the driver can indicate
when the change in channel status actually took place.

Related to this, clarify in a comment in ring_buffer.fidl
that the `set_time` parameter returned in response to
`SetActiveChannels` can be a timestamp before the method
call altogether, if the `SetActiveChannels` call represented
no change to the hardware configuration.

This CL also enables `SetActiveChannels` for VirtualAudio
Composite, Dai and StreamConfig drivers.

Test: fx test -ouv audio_drivers_tests -- --admin

Change-Id: Ic2e187bbee4e515aabe5639cab4f22e5734c9431
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961875
Reviewed-by: Andres Oportus <andresoportus@google.com>
API-Review: Marie Janssen <jamuraa@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>