[display] Replace "G101B158" with "P101DEZ".
The model of the Innolux panel used on the Sherlock board was incorrect.
The U-boot code has the correct model. This CL aligns the Fuchsia code
with the U-boot code on the model, to avoid any future confusion.

Test: fx build (no functional change)
Change-Id: I4c84dfc2735e42a334dedd3577831f8f0bd63c20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973583
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Victor Costan <costan@google.com>