[fxfs] Fix an issue with continuing after an unclean unmount
This bug meant that Fxfs would resume writing to the journal with the
wrong checksum after an unclean unmount.  This would mean that no more
writes would persist and could lead to corruption if existing data was
deallocated.

The test also uncovered another issue which is that when discarding a
large part of the journal, the space for it would come out of the next
transaction's reservation which might not be big enough to accommodate
it. This fixes that by borrowing that space if it is significant.

Bug: b/319293601
Change-Id: I00a7af12e50dd366757d4b7e324e87853c226149
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973764
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>