[starnix] Implement fusectl filesystem
This CL adds the fusectl filesystem and handles the abort and waiting
control file in it.

It also changed the error returned when a fuse filesystem is
disconnected to ECONNABORTED to ensure the error is fatal.

Fixed: b/319836413, b/317211006
Change-Id: I35d049b47d055626329d83afb5ec44ec756605d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974312
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>