[bazel] Use `config` from assembly_bundle
... instead of `dir`. In general Bazel discourages using directories as
inputs and outputs, which will result in unsound dependency checks. In
Bazel 7 such usage will result in build errors by default [0].

In `assembly_bundle`, the `dir` field is necessary to locate the
assembly bundle directory. The same can be achieved by pointing to an
actual file in the root of the directory (assembly_config.json should
always exist in AIBs), and get the path to its parent directory.

Also removed `platform_aib.bzl`, this rule currently can only output a
directory, and it's non-trivial to make it list all files from the
directory in outputs. One common alternative is to create an archive
from this directory and use that as the output. This rule is not needed
until we are ready to start migrating AIB definitions from GN to Bazel,
and we can revisit when we are ready to do that.

[0] https://github.com/bazelbuild/bazel/issues/18646

Bug: 319069000

Change-Id: I783dc5a2d2a3ee69ec0b97e670ad4f105ae07b4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973216
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Turner <digit@google.com>