[fxfs] Fix a corruption race.
`flush_all_files()` is called from a background thread where it
could race against `open_count_sub_one()` leading to flushing of a file that is in the process of being purged. This can sometimes lead to bad data. (ObjectValue::None when updating allocated size.)

This wraps flush() so that we never try to flush a file marked as TO_BE_PURGED.

Bug: b/318773519
Change-Id: Ie6a6f61e25ded848304c0567304c985bc4cefe69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973214
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>