[perf_publish] Unify catapult path and expected metrics path
These two arguments are only used in tests at the moment. We can unify
them into a single one that also reflects the typical situation, in
which both the binary and the file are placed in the runtime_deps
directory of the test.

Bug: 305082234
Bug: 305081962
Change-Id: I5a90a4bf7fecc6b8bd456e3ace3ffddf523a2427
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972653
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>