[fidlc] Remove nested namespaces
This CL removes all nested namespaces in fidlc to comply with
https://fuchsia.dev/fuchsia-src/development/languages/c-cpp/cpp-style#namespace_names.
It also changes the top-level namespace from "fidl" to "fidlc".

C++ namespaces are meant for avoiding name collisions between projects.
They are not very well suited for organizing code within a project.
See https://abseil.io/tips/130 for more information on the pitfalls of
nested namespaces.

The fidlc codebase is small enough that it's no trouble to come up with
unique names. When names do collide, it's better to come up with more
descriptive names for both, and not need to worry about the current
namespace scope to know what a name refers to.

The main place where we relied on nested namespaces to avoid collisions
was the raw/flat AST. I removed these namespaces and prefixed raw AST
types with "Raw" instead. (This is less typing than "raw::".)

A follow-up CL will remove the flat/ subdirectory now that it no longer
corresponds to a namespace.

Test: fx test fidlc-test
Change-Id: Id4482b503050fcc77d01c072644415f7bdb5266f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973600
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ian McKellar <ianloic@google.com>