[fsverity] Fail writes once enable_verity is called
The first thing we do when enabling verity is call flush() which cleans
all the pages in the file's vmo. In order to ensure that no pages are
written to after the flush, we set a bit at the top of enable_verity and
fail all calls to mark_dirty after that bit is set for that file.

Bug: b/316935310
Bug: b/314195208
Change-Id: I7730dffaac52648d74e435bfa5580a1e22eb67bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969693
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Nikita Jindal <nikitajindal@google.com>