[component_manager] Fix namespace shutdown
Right now, when a component stops, the namespace doesn't get shutdown
and their associated tasks owned by the registry continue to run. This
regression was introduced at some point during the bedrock introduction.

To fix this, call [ExecutionScope::shutdown] on the namespace scope
as part of the stopping procedure.

Now, each Directory capability does not create its own scope, but
accepts one as an argument. It's now up to the client to pass a scope.

There is apparently still a bug lurking because the tasks are not
getting dropped on their own, but nevertheless this is worth doing
because it a) makes shutdown explicit, b) unifies the namespace under a
single scope, c) paves the way for graceful namespace shutdown
(fxbug.dev/303919602).

Bug: 319780352, 303919602

Change-Id: I002422c6137197125520e56e76b67b1657b44517
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973592
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Gary Bressler <geb@google.com>