[network/tests] Log interfaces watcher events
... for delegated_provisioning_test.

This test waits to observe an address being assigned on a specific
interface by polling the interfaces watcher. This test has recently
flaked, during which the logs indicate that the address was added but
that the test did not observe the change.

Add logging so that if the flake re-occurs, we can deduce which
component is at fault.

Bug: b/320263602
Change-Id: I2a81c7bfb5def74a9fa52edbf84274afdc1e7479
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975352
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>