[test] test_spec must point to target-under-test
fuchsia_test_package was using test_spec to declare a test target, but
the test_spec was pointing to itself as a target instead of the
package-under-test. This was resulting in these tests not being
"affected", because they did not depend on anything.

Change-Id: I5b1de5dd58451049b9108a0b6fbcc51f69ba2794
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975573
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>