[starnix] Support for ptrace stop when cloning
Includes the following changes to ptrace-event-stop behavior:

1) When ptrace waits for a ptrace-event-stop, the type of event-stop
was previously (incorrectly) stored in the signal's code and made
visible to the user.  It is now stored in the ExitStatus for Stop and
Continue.

2) PTRACE_EVENTs now have an enum.

3) PTRACE_EVENTMSG is now implemented: it provides the user a word
containing information about the type of event.  In the case of clone,
it provides the tid of the cloned task. Information about the event
is stored in a new struct: PtraceEventData.

4) The data that needs to be propagated from parent to cloned child
in order to attach to the same tracer is extracted into its own
struct: PtraceCoreData.

This CL adds three lock acquisition / releases to the clone operation:
two in the parent, one in the child.  They take place to check if the
operation is supposed to be traced.

Change-Id: I6897225af37ad6cd7a434970b711fc0efab44bd3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956232
Commit-Queue: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Adam Barth <abarth@google.com>