[tracing] Add component to stress tracing with many threads
This component spawns a significant number of threads that actively
write to the same trace provider, and hence the same trace buffer, for
the purpose of stressing the tracing system and exposing various race
conditions.

Additionally, this patch adds a collection under trace_manager for the
purpose of running components that are primarily focused at producing a
workload for the tracing system.

BUG: b/303666620

Change-Id: I044148948a4dd8ebaa1b62ce7b9d68ca622b835b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938819
Commit-Queue: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Krithika Balan <krithikabalan@google.com>