[wlan][drivers] Add TRACE_INSTANT() in every WLAN_TRACE_DURATION_()
When a crash occurs in the same scope as a TRACE_DURATION() macro
call, the duration does not appear in the generated trace. This
hinders debugging since one of the durations you're most interested in
during a crash is the one in scope at the time of the crash.

This change adds a TRACE_INSTANT to every call to
WLAN_TRACE_DURATION_() so that even in the event of a crash, at least
the instant will appear in the trace.

Change-Id: Icc5e106b6bb9d71d7c993e3383e50f9f7dd9c9e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974812
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kiet Tran <kiettran@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>