[netcfg] Improve bus type matching
SDIO is a particular bus for doing IO over a Secure Digital card
interface. It's not quite appropriate to classify every platform bus as
SDIO; the controller could be on-chip, virtualized, or even attached to
some low-bandwidth bus like i2c.

We already discriminate PCI and USB from SDIO, so we probably ought to
be more careful in admitting when we don't know which bus carries the
device. This also reduces reliance on topological path, which ought to
be fine: if we discover a network device, we can still provision it
without knowing how it's attached.

This change introduces two new bus types: "Virtio", which uses standard
bus path naming rules, and "Unknown", which uses MAC-address-based
naming. This also removes a few tests that depended on bus path and bus
type matching failures as these can no longer fail.

Bug: b/319153091
Change-Id: Ibc42a19f99d7f900bf2ae494cae2a0416d93a97d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976032
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>