[tracing] Remove the Buffer Size Limit
Originally intended as a fail safe "no one should need more than 64MB"
we do have use cases where we use larger buffer sizes. Since we don't
actually map or allocate the memory until it's used, we should be okay
to allow larger buffer sizes to be specified.

In addition, remove the intermediate buffer that we copy the full trace
buffer to regardless of size. This was very slow to do when we have many
trace providers. Instead, map the whole buffer and skip the copy.

Fixed: 321082169
Change-Id: I17c28d6451767242bffb30224489b32fc3c8ced0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976540
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Krithika Balan <krithikabalan@google.com>