[superproject] Roll cobalt Create a new log source enum proto.
This is the first step towards using the internal copies when building
Cobalt Core internally. The protos need to use enums, as that is how
they are defined internally. The new enum is not used yet, but I need
to create the new file so I can add it to the internal build rules,
so as to not break the internal roller in the next change.

Original-Bug: b/319167264
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/970978
Original-Revision: 8adfed2ae914ca2e50a4d7161ff4ff7715b84c0d
GitOrigin-RevId: 6f2c9db7c74a91adfcdd4425d5b740bd2780cfdb
Roller-URL: https://ci.chromium.org/b/8758169149155296977
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0c2a3c8371c9c8d6124e9ace3d1ca5c0dadd008b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978392