[rust-3p] Add helper to generate bitflags 1.x compatible serde traits
bitflags 2.x changed the serde serialization format, which broke parsing
the legacy format. This adds a new helper library that supports
serializing and deserializing from the 1.x format, as well as
deserializing from the 2.x format. This can eventually be used as a way
to gracefully migrate the ecosystem to the new format.

Bug: b/299339125
Change-Id: Ia3674ad2f2f15a42d16fb41307cbb9520c1a8f93
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978593
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>