[vim3] more contig ram + enable page loaning
On vim3, framebuffers are physically contiguous, which requires a
significant contiguous_memory_size. To avoid stranding pages when
they're not used for framebuffers (and other contiguous buffers),
enable page loaning so that zircon can use these pages in pager-backed
VMOs.

This also increases contiguous_memory_size to 384 MiB, but any unused
portion of that gets decommitted back to zircon.

The "k ppb stats" command can be used to check how much is loaned back
to zircon, and how much of that is used in pager-backed VMOs.

BUG: b/42072489
BUG: b/322009732
Change-Id: Ia33ff46761abb0ab94ebf112427d74f72aed44e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975806
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>