[superproject] Roll cobalt Standardize the clearcut protos to look more like our internal copies.
This is the 3rd step towards using the internal copies when building
Cobalt Core internally.

The proto packages are all made consistent to a new package name that
can be replaced with the internal name by our roller. Some fields that
we used to set as integers are actually enums, so I created enums for
them. All these changes should be compatible, resulting in the same
binary protos as we currently generate.

Original-Bug: b/319167264

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/970977
Original-Revision: 4a55bb3ac357185e0e396483fb4e5342e1ea8030
GitOrigin-RevId: a6506bdb28fb7ac69221ce8ecd9b9d591efd40d6
Roller-URL: https://ci.chromium.org/b/8757987467542119601
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0b1bd646c2dbf22e4ec6fe5fce38e7dd6fe85374
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979735