[recovery] Don't fall back to wiping Fxfs
Previously, if the request to wipe the data volume from fshost failed,
we would fall back to manually erasing the Fxfs partition by overwriting
the superblocks.  This was sensible when Fxfs just stored mutable data,
but now Fxfs also stores blobs, so this will brick the device.

Remove the fallback logic.  This required plumbing the fxfs_blob setting
from the build into the factory_reset component, so that we would know
whether the fshost failure should be bubbled up (if fxblob is enabled)
or to continue by trying to wipe zxcrypt (if disabled).

Note that this removes the fallback logic for the Fxfs-but-not-fxblob
case too.  I didn't think it was worth also supporting that case, but if
we felt that it was important enough, we could keep it.

Change-Id: I6d8b062f24449426bdbc2f44f894a1befa0c4ee1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975104
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>